Create separate file for Manifest trait helpers
- Move the trait helper and validation methods to a separate file
  (Manifest/Manifest+Traits.swift) and move the TraitConfiguration
  into the PackageModel instead of PackageGraph; updated the
  appropriate CMakeLists.txt
- Update TraitConfiguration computed variable for enabled traits;
  '.none' case should return '["default"]', as that is what its
   behaviour defaults to. Removed other computed variables in
  favor of checking the enum cases themselves where appropriate.
