+++ Sources/Commands/PackageCommands/Resolve.swift
@@ -14,7 +14,7 @@ import ArgumentParser
 import CoreCommands
 import TSCUtility
 
-import enum PackageGraph.TraitConfiguration
+import enum PackageModel.TraitConfiguration
 
 extension SwiftPackageCommand {
     struct ResolveOptions: ParsableArguments {
